Another One
For this commit I added some elements such as, a picture (logo for the website) as a test for the ease-in-out element and more portfolio pieces on the page. What I would like to be looked at would be, to get rid of the boxes being cut off by the margin, some elements that are not working like the other ones such as the typeswap poster box and the modular typeface box.