3/21/24
Made some positional changes to the overall design of the website. Please look at the background of the overall website, we talked about a hatch kind of style for a manga. Also, I am going to look into how I can get some styles so we can make it really look like a manga. Thanks, have a good weekend.