4/16/2024
Added some pictures to help make the positioning of things a lot easier. Added a background that needs to be positioned correctly in order to make the background appear clear but also faded so it isn't the main focus of the website. Some other positioning changes for the smaller sizes of the website so the pictures and headings are not cluttered on the website. Trying to get some illustration in for the website maybe some placeholders possibly.